[improvement, fix] - Distribute structs, fix circular dependency and compilation order issues
