+++ modules/rendering/include/gltf-scene-importer.h
@@ -24,7 +24,7 @@
 #pragma once
 
 #include <GL/glew.h>
-#include "iscene-importer.h"  
+#include <iscene-importer.h>  
 #include <string>
 #include <stdexcept>
 #include <iostream>
