+++ modules/rendering/src/opengl-shader.cpp
@@ -1,6 +1,5 @@
 #include <opengl-shader.h>
 
-
 /**
  * @brief Constructor for the OpenGlShader class.
  *
